taskflow-api/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
alembic
psycopg[binary]
pydantic>=2
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
email-validator
redis
slowapi
prometheus-fastapi-instrumentator
asgi-correlation-id
structlog
python-json-logger
httpx
pytest
pytest-asyncio
docker
python-multipart
fastapi-cache2
redis[hiredis]